+++ models/User.go
@@ -13,6 +13,7 @@ type User struct {
 	FirstName  string `form:"first_name" gorm:"not null; type:varchar(55)"`
 	Email      string `form:"email" gorm:"unique; not null; type:varchar(55)"`
 	Password   string `form:"password" gorm:"not null; type:varchar(255)"`
+	Admin      bool   `form:"admin" gorm:"default:0; type:boolean"`
 }
 
 // UserGet defines the structure of the user API response entity
@@ -21,6 +22,7 @@ type UserGet struct {
 	LastName  string    `json:"last_name"`
 	FirstName string    `json:"first_name"`
 	Email     string    `json:"email"`
+	Admin     bool      `json:"admin"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
